pyproject.toml
[project]
name = "dnd-hackathon"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord>=2.3.2",
    "dotenv>=0.9.9",
    "honcho-ai>=1.4.0",
]
